Tie together various sections of Chess game
Implement class to bring all the functionality of the various pieces
and the chessboard to play a game of Chess. Implement class to drive
the running of one or more games of Chess as well as manage saved Chess
games.
Implement specs for new methods and Rakefile to run application